Extract gear card to component
