+++ packages/web/src/features/profile/components/Gear.tsx
@@ -1,28 +1,18 @@
-import { FC, useCallback, useEffect, useMemo, useState } from 'react';
-import { Button, HStack, Image, Stack, Select } from '@chakra-ui/react';
+import { FC, useMemo } from 'react';
+import { Button, HStack } from '@chakra-ui/react';
 import { useWeb3React } from '@web3-react/core';
-import { utils } from 'ethers';
-import PanelBody from 'components/PanelBody';
 import {
   Item,
   Maybe,
-  useHustlersWalletQuery,
   useInfiniteProfileGearQuery,
   WalletItems,
 } from 'generated/graphql';
-
-import { css } from '@emotion/react';
+import CardContainer from './CardContainer';
+import GearCard from './GearCard';
 import ItemCount from './ItemCount';
-import ProfileCard from './ProfileCard';
-import ProfileCardHeader from './ProfileCardHeader';
+import LoadingBlock from 'components/LoadingBlock';
 import SectionContent from './SectionContent';
 import SectionHeader from './SectionHeader';
-import CardContainer from './CardContainer';
-import LoadingBlock from 'components/LoadingBlock';
-import PanelFooter from 'components/PanelFooter';
-import { useSwapMeet } from 'hooks/contracts';
-import { NETWORK } from 'utils/constants';
-import { useOptimism } from 'hooks/web3';
 
 type ProfileItem = Pick<WalletItems, 'id' | 'balance'> & {
   item: Pick<Item, 'id' | 'count' | 'fullname' | 'name' | 'svg' | 'suffix' | 'type'> & {
@@ -33,80 +23,6 @@ type ProfileItem = Pick<WalletItems, 'id' | 'balance'> & {
 type GearData = {
   walletItems: ProfileItem[];
   totalCount: number;
-};
-
-const getOrigin = (suffix?: string | null): string => {
-  if (!suffix) return '...';
-
-  const [, origin] = suffix.split('from ');
-
-  return origin;
-};
-
-const getImageSrc = (walletItem: ProfileItem): string => {
-  return walletItem.item.svg || walletItem.item.base?.svg || '';
-};
-
-const GearFooter = ({ id }: { id: string }) => {
-  const { account } = useWeb3React();
-  const { chainId } = useOptimism();
-  const [selected, setSelected] = useState<string>();
-  const { data, isFetching: loading } = useHustlersWalletQuery(
-    {
-      where: {
-        id: account,
-      },
-    },
-    {
-      enabled: !!account,
-    },
-  );
-  const swapmeet = useSwapMeet();
-
-  useEffect(() => {
-    if (
-      data &&
-      data.wallets.edges &&
-      data.wallets.edges[0] &&
-      data.wallets.edges[0].node &&
-      data.wallets.edges[0].node.hustlers.length > 0
-    ) {
-      setSelected(data.wallets.edges[0].node.hustlers[0].id);
-    }
-  }, [data]);
-
-  const equip = useCallback(() => {
-    const sig = '0xbe3d1e89';
-    const abi = new utils.AbiCoder();
-    swapmeet.safeTransferFrom(
-      account!,
-      NETWORK[chainId].contracts.hustlers,
-      id,
-      1,
-      abi.encode(['bytes4', 'uint256'], [sig, selected]),
-    );
-  }, [account, swapmeet, chainId, id, selected]);
-
-  return (
-    <PanelFooter>
-      <Select
-        size="sm"
-        variant="filterBar"
-        onChange={({ target }) => setSelected(target.value)}
-        value={selected}
-      >
-        <option disabled>Equip to…</option>
-        {data?.wallets.edges![0]?.node?.hustlers.map(({ id, title, name }) => (
-          <option key={id} value={id}>
-            {title} {name}
-          </option>
-        ))}
-      </Select>
-      <Button variant="primary" disabled={!selected} onClick={equip}>
-        Equip
-      </Button>
-    </PanelFooter>
-  );
 };
 
 const GearWrapper: FC = () => {
@@ -171,39 +87,13 @@ const GearWrapper: FC = () => {
         {gearData.walletItems.length ? (
           <CardContainer>
             {gearData.walletItems.map(walletItem => {
-              const origin = getOrigin(walletItem.item.suffix);
-              const imageSrc = getImageSrc(walletItem);
-
               return (
-                <ProfileCard key={walletItem.item.id}>
-                  <ProfileCardHeader>
-                    <div>{walletItem.item.name}</div>
-                    <div
-                      css={css`
-                        padding-right: 16px;
-                        color: var(--gray-500);
-                      `}
-                      title="You have this many in inventory"
-                    >
-                      x{walletItem.balance}
-                    </div>
-                  </ProfileCardHeader>
-                  <PanelBody>
-                    <Stack>
-                      <Image borderRadius="md" src={imageSrc} alt={walletItem.item.name} />
-                      <span>Type: {walletItem.item.type}</span>
-                      <span>Origin: {origin}</span>
-                      <span
-                        css={css`
-                          height: 2.5em;
-                        `}
-                      >
-                        Title: {walletItem.item.fullname}
-                      </span>
-                    </Stack>
-                  </PanelBody>
-                  <GearFooter id={walletItem.item.id} />
-                </ProfileCard>
+                <GearCard 
+                  key={walletItem.id} 
+                  item={walletItem.item} 
+                  balance={walletItem.balance} 
+                  showEquipFooter
+                />
               );
             })}
             {isFetching && gearData.walletItems.length && <LoadingBlock maxRows={1} />}
